Moved method "convert" to presenter
MVC -> MVP
